requirements.txt
# frame requirements
dynaconf
flask
flask-cors
# model requirements
numpy==1.21.2
opencv-python==4.5.5.64
pyyaml==6.0
torch==1.11.0
pillow==8.4.0
torchvision==0.12.0
Cython
matplotlib==3.5.1
torchsummary==1.5.1